requirements.txt
tensorflow-cpu==2.11.0
librosa
numpy
pandas
scikit-learn
matplotlib
streamlit
